added files skeleton, moved sidebar to base template with blocks for inserting other templates' info.
