+++ jt/templates/base.html
@@ -16,8 +16,37 @@
     {% include 'navbar.html'  %}
 
     <div class="main-container">
-    {%block content%}
-    {%endblock%}
+      <div class="columns">
+
+        <div class="column is-one-quarter column1 has-background-light">
+          <div class="col-1-div">
+            <h2 class="categories-header">CATEGORIES</h2>
+            <ul class="categories-list">
+              {% for category in category_list %}
+              <li><a href="#" class="category">{% category.name %}</a></li>
+              {% endfor %}
+              <li><a href="#" class="category">Animals</a></li>
+              <li><a href="#" class="category">Knock Knock</a></li>
+              <li>etc.</li>
+            </ul>
+          </div>
+        </div>
+
+        <div class="column is-half column2">
+          <div class="col-2-div">
+            {%block center%}
+            {%endblock%}
+          </div>
+        </div>
+
+        <div class="column is-one-quarter column3">
+          <div class="col-3-div">
+            {%block right%}
+            {%endblock%}
+          </div>
+        </div>
+
+      </div>
     </div>
     
   </body>
