+++ jt/templates/index.html
@@ -4,33 +4,17 @@
 <link rel="stylesheet" type="text/css" href="{% static 'jt/index.css' %}">
 <link rel="stylesheet" href="{% static 'jt/node_modules/bulma/css/bulma.css' %}">
 
-{%block content%}
-  <div class="columns">
-
-    <div class="column is-one-quarter column1 has-background-light">
-      <div class="col-1-div">
-        <h2 class="categories-header">CATEGORIES</h2>
-        <ul class="categories-list">
-          <li><a href="#" class="category">Animals</a></li>
-          <li><a href="#" class="category">Knock Knock</a></li>
-          <li>etc.</li>
-        </ul>
-      </div>
+{%block center%}
+  <div class="flip3D">
+    <div class="back">
+      <h3 class="card-text">Because they have such big fingers!</h3>
     </div>
-    <div class="column is-half column2">
-      <div class="col-2-div">
-        <div class="flip3D">
-          <div class="back">
-            <h3 class="card-text">Because they have such big fingers!</h3>
-          </div>
-          <div class="front">
-            <h3 class="card-text">Why do gorillas have such big nostrils?</h3>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="column is-one-quarter column3">
-      3
+    <div class="front">
+      <h3 class="card-text">Why do gorillas have such big nostrils?</h3>
     </div>
   </div>
+{%endblock%}
+
+{%block right%}
+  <h1>Hello World!</h1>
 {%endblock%}
