+++ jt/views/__init__.py
@@ -1,0 +1,2 @@
+from .views import index
+from .views import joke
